Add option to specify maximum number of threads for multi-start optimization
So far, this was equal to the number of local optimizations, but that wouldn't scale well.

This decouples the number parallel local optimizations from the total number of optimizations.

This is controlled by setting environment variable `PARPE_NUM_PARALLEL_STARTS` to the number
of optimizations that should be run in parallel.

Now requires Boost.Thread library

Closes #359